requirements.txt
numpy == 1.22.3
pandas == 1.4.2
matplotlib == 3.5.1
Jinja2 == 3.0.3
Flask == 2.1.2
scipy == 1.8.0
scikit-learn == 1.1.1
xgboost == 1.6.0
seaborn==0.12.0
autoimpute==0.13.0
pydotplus==2.0.2
ipykernel==6.23.0
cassandra-driver==3.25.0
cassandra-csv==0.3
astrapy==0.0.2
simplejson==3.17.2
mysql==0.0.3
pymongo==4.3.3
-e .
